sif: move ACPI root-bus discovery and IRQ routing into pci/acpi.rs
